requirements.txt
python_telegram_bot==20.1
python-dotenv
molmass[all]
unidecode
sqlalchemy